feat: Create additional filter and saving recents
Additional filter is a set of features that change filtering method or
add another filter. This can be shown or hidden by clicking side button,
which is going to be fixed to different design. For now, only two
features are made, one called "Auto Start" is to start automatically
when selection is finished, the other called "Exclude Recent Music" is
to adjust how many songs selected recently will be ruled out. Favorite
is not created yet.

Saving recents is to save list of songs selected even when this
application is closed.

Minor updates
- Track list won't be updated when track list setting is not changed.