feat: Add mode and level options to add-on
Online mode just selects music, not difficulty. History will show that
selected music's style is "FREE."
Beginner and Master level selects pattern which has easiest or hardest
difficulty in selected music.

When Online mode is selected, Auto Start in Aider and all features in
Level(add-on) will not work, because these features are unnecessary for
in-game online match.